+++ oidc-idp/src/main/java/cz/muni/ics/oidc/server/filters/impl/PerunIsCesnetEligibleFilter.java
@@ -1,0 +1,149 @@
+package cz.muni.ics.oidc.server.filters.impl;
+
+import cz.muni.ics.oidc.BeanUtil;
+import cz.muni.ics.oidc.models.PerunAttribute;
+import cz.muni.ics.oidc.models.PerunUser;
+import cz.muni.ics.oidc.server.configurations.PerunOidcConfig;
+import cz.muni.ics.oidc.server.connectors.PerunConnector;
+import cz.muni.ics.oidc.server.filters.FiltersUtils;
+import cz.muni.ics.oidc.server.filters.PerunFilterConstants;
+import cz.muni.ics.oidc.server.filters.PerunRequestFilter;
+import cz.muni.ics.oidc.server.filters.PerunRequestFilterParams;
+import cz.muni.ics.oidc.web.controllers.ControllerUtils;
+import cz.muni.ics.oidc.web.controllers.PerunUnapprovedController;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+
+import static cz.muni.ics.oidc.server.filters.PerunFilterConstants.PARAM_FORCE_AUTHN;
+import static cz.muni.ics.oidc.server.filters.PerunFilterConstants.PARAM_REASON;
+import static cz.muni.ics.oidc.server.filters.PerunFilterConstants.PARAM_SCOPE;
+import static cz.muni.ics.oidc.server.filters.PerunFilterConstants.PARAM_TARGET;
+import static cz.muni.ics.oidc.web.controllers.PerunUnapprovedController.REASON_EXPIRED;
+import static cz.muni.ics.oidc.web.controllers.PerunUnapprovedController.REASON_NOT_SET;
+
+/**
+ * This filter verifies that user attribute isCesnetEligible is not older than given time frame.
+ * In case the value is older, denies access to the service and forces user to use verified identity.
+ * Otherwise, user can to access the service.
+ *
+ * Configuration (replace "name" part with a name defined for the filter):
+ * - filter.name.isCesnetEligibleAttr - mapping to isCesnetEligible attribute
+ * - filter.name.validityPeriod - specify in months, how long the value can be old, if no value or invalid value
+ * provided, defaults to 12 months
+ *
+ * @author Dominik Frantisek Bucik <bucik@ics.muni.cz>
+ */
+public class PerunIsCesnetEligibleFilter extends PerunRequestFilter {
+
+    private static final Logger log = LoggerFactory.getLogger(PerunIsCesnetEligibleFilter.class);
+
+    /* CONFIGURATION PROPERTIES */
+    private static final String IS_CESNET_ELIGIBLE_ATTR_NAME = "isCesnetEligibleAttr";
+    private static final String IS_CESNET_ELIGIBLE_SCOPE = "isCesnetEligibleScope";
+    private static final String VALIDITY_PERIOD = "validityPeriod";
+    private static final String DATE_TIME_FORMAT = "yyyy-MM-dd HH:mm:ss";
+
+    private final String isCesnetEligibleAttrName;
+    private final String isCesnetEligibleScope;
+    private final int validityPeriod;
+    /* END OF CONFIGURATION PROPERTIES */
+
+    private final PerunConnector perunConnector;
+    private final PerunOidcConfig perunOidcConfig;
+
+    public PerunIsCesnetEligibleFilter(PerunRequestFilterParams params) {
+        super(params);
+
+        BeanUtil beanUtil = params.getBeanUtil();
+
+        this.perunConnector = beanUtil.getBean(PerunConnector.class);
+        this.perunOidcConfig = beanUtil.getBean(PerunOidcConfig.class);
+
+        this.isCesnetEligibleAttrName = params.getProperty(IS_CESNET_ELIGIBLE_ATTR_NAME);
+        this.isCesnetEligibleScope = params.getProperty(IS_CESNET_ELIGIBLE_SCOPE);
+        int validityPeriodParam = 12;
+        if (params.hasProperty(VALIDITY_PERIOD)) {
+            try {
+                validityPeriodParam = Integer.parseInt(params.getProperty(VALIDITY_PERIOD));
+            } catch (NumberFormatException ignored) {
+                //no problem, we have default value
+            }
+        }
+
+        this.validityPeriod = validityPeriodParam;
+    }
+
+    @Override
+    protected boolean process(ServletRequest req, ServletResponse res) {
+        HttpServletRequest request = (HttpServletRequest) req;
+        HttpServletResponse response = (HttpServletResponse) res;
+
+        if (!FiltersUtils.isScopePresent(request.getParameter(PARAM_SCOPE), isCesnetEligibleScope)) {
+            log.debug("Scope {} not present in request, no point in continuing filter execution. Skip to the next filter",
+                    isCesnetEligibleScope);
+            return true;
+        }
+
+        PerunUser user = FiltersUtils.getPerunUser(request, perunOidcConfig, perunConnector);
+        if (user == null) {
+            log.warn("Could not extract user from request, skip to the next filter");
+            return true;
+        }
+
+        String reason;
+
+        PerunAttribute attribute = perunConnector.getUserAttribute(user.getId(), isCesnetEligibleAttrName);
+        if (attribute != null && attribute.getValue() != null && attribute.valueAsString() != null) {
+            LocalDateTime timeStamp;
+            try {
+                DateTimeFormatter formatter = DateTimeFormatter.ofPattern(DATE_TIME_FORMAT);
+                timeStamp = LocalDateTime.parse(attribute.valueAsString(), formatter);
+            } catch (DateTimeParseException e) {
+                log.error("Could not parse {} value: {}", isCesnetEligibleAttrName, attribute.valueAsString());
+                return true;
+            }
+
+            LocalDateTime now = LocalDateTime.now();
+            if (now.minusMonths(validityPeriod).isBefore(timeStamp)) {
+                log.debug("{} valid, go to the next filter", isCesnetEligibleAttrName);
+                return true;
+            } else {
+                reason = REASON_EXPIRED;
+            }
+        } else {
+            reason = REASON_NOT_SET;
+        }
+
+        log.debug("Value of the attribute is invalid, redirecting to unauthorized");
+        redirect(request, response, reason);
+
+        return false;
+    }
+
+    private void redirect(HttpServletRequest req, HttpServletResponse res, String reason) {
+        Map<String, String> params = new HashMap<>();
+
+        String targetURL = FiltersUtils.buildRequestURL(req,
+                Collections.singletonMap(PARAM_FORCE_AUTHN, "true"));
+        params.put(PARAM_TARGET, targetURL);
+        params.put(PARAM_REASON, reason);
+
+        String redirectUrl = ControllerUtils.createRedirectUrl(req, PerunFilterConstants.AUTHORIZE_REQ_PATTERN,
+                PerunUnapprovedController.UNAPPROVED_IS_CESNET_ELIGIBLE_MAPPING, params);
+        res.reset();
+        res.setStatus(HttpServletResponse.SC_MOVED_PERMANENTLY);
+        res.setHeader("Location", redirectUrl);
+    }
+
+}